requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
torch>=1.12.0
timm>=0.9.0
torchvision
numpy
pillow
opencv-python-headless
scikit-image
scipy
